+++ src/test/java/nextstep/common/acceptance/AcceptanceTest.java
@@ -1,6 +1,6 @@
-package nextstep.subway.acceptance;
+package nextstep.common.acceptance;
 
-import nextstep.subway.utils.DatabaseCleanup;
+import nextstep.common.utils.DatabaseCleanup;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
