+++ src/test/java/nextstep/common/utils/DatabaseCleanup.java
@@ -1,4 +1,4 @@
-package nextstep.subway.utils;
+package nextstep.common.utils;
 
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.context.annotation.Profile;
